% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_IBD.R
\name{Check_IBD}
\alias{Check_IBD}
\title{Check properties of an incomplete block design (IBD)}
\usage{
Check_IBD(Design)
}
\arguments{
\item{Design}{Provide an IBD in matrix format}
}
\value{
Provides C matrix (Information matrix), eigenvalues(EVs) and canonical efficiency factor (CEF) of a given IBD
}
\description{
Check properties of an incomplete block design (IBD)
}
\examples{
library(SudokuDesigns)
Design<-matrix(c(1,2,3,2,5,3,2,4,6),nrow=3,byrow=TRUE)
Check_IBD(Design)
}
