% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_IRC.R
\name{Check_IRC}
\alias{Check_IRC}
\title{Check properties of an incomplete row-column design (IRC)}
\usage{
Check_IRC(Design)
}
\arguments{
\item{Design}{Provide an IRC in matrix format}
}
\value{
Provides C matrix (Information matrix), eigenvalues(EVs) and canonical efficiency factor (CEF) of a given IBD
}
\description{
Check properties of an incomplete row-column design (IRC)
}
\examples{
library(SudokuDesigns)
Design<-matrix(c(1,2,3,2,5,3,2,4,6),nrow=3,byrow=TRUE)
Check_IRC(Design)
}
