% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_MP_Inverse.R
\name{Check_MP_Inverse}
\alias{Check_MP_Inverse}
\title{Moore Penrose Inverse}
\usage{
Check_MP_Inverse(matrix)
}
\arguments{
\item{matrix}{Any matrix}
}
\value{
Provides Moore Penrose inverse of a given matrix
}
\description{
Moore Penrose Inverse
}
\examples{
library(SudokuDesigns)
mat<-matrix(c(1,2,3,2,5,3,2,4,6),nrow=3,byrow=TRUE)
Check_MP_Inverse(mat)
}
