% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Obsn_vs_Reg_Matrix.R
\name{Check_Obsn_vs_Reg_Matrix}
\alias{Check_Obsn_vs_Reg_Matrix}
\title{Observations Vs Regions Incidence Matrix}
\usage{
Check_Obsn_vs_Reg_Matrix(Design, Region)
}
\arguments{
\item{Design}{A Sudoku design in matrix format}

\item{Region}{A matrix of regions according to the Sudoku design}
}
\value{
Observations vs regions incidence matrix for a given Sudoku design and region matrix
}
\description{
Observations Vs Regions Incidence Matrix
}
\examples{
library(SudokuDesigns)
design<-matrix(c(1,2,3,4,3,4,1,2,2,1,4,3,4,3,2,1),nrow=4,ncol=4,byrow=TRUE)
region<-matrix(c(1,1,2,2,1,1,2,2,3,3,4,4,3,3,4,4),nrow=4,ncol=4,byrow=TRUE)
Check_Obsn_vs_Reg_Matrix(design, region)
}
