% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Obsn_vs_Trt_Matrix.R
\name{Check_Obsn_vs_Trt_Matrix}
\alias{Check_Obsn_vs_Trt_Matrix}
\title{Observations Vs Treatments Incidence Matrix}
\usage{
Check_Obsn_vs_Trt_Matrix(Matrix)
}
\arguments{
\item{Matrix}{Any matrix}
}
\value{
Generates observations Vs treatments matrix
}
\description{
Observations Vs Treatments Incidence Matrix
}
\examples{
library(SudokuDesigns)
mat1<-matrix(c(1,2,3,4,1,3,6,2,8,1,8,3),nrow=4,byrow=TRUE)
mat1
Check_Obsn_vs_Trt_Matrix(mat1)
}
