% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Rank.R
\name{Check_Rank}
\alias{Check_Rank}
\title{Checking Rank of a Matrix}
\usage{
Check_Rank(matrix)
}
\arguments{
\item{matrix}{Any matrix}
}
\value{
Print the rank of the given matrix
}
\description{
Checking Rank of a Matrix
}
\examples{
library(SudokuDesigns)
mat<-matrix(c(1,2,3,2,4,6,5,2,3),nrow=3,byrow=TRUE)
Check_Rank(mat)
}
