% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Replication_Matrix.R
\name{Check_Replication_Matrix}
\alias{Check_Replication_Matrix}
\title{Replication Matrix}
\usage{
Check_Replication_Matrix(matrix)
}
\arguments{
\item{matrix}{Any matrix}
}
\value{
A diagonal matrix of replications for each treatment.
}
\description{
Replication Matrix
}
\examples{
library(SudokuDesigns)
mat11<-matrix(c(1,2,3,4,1,3,6,2,8,1,8,3),nrow=4,byrow=TRUE)
mat11
Check_Replication_Matrix(mat11)
}
