% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Replications.R
\name{Check_Replications}
\alias{Check_Replications}
\title{Replications for each treatments}
\usage{
Check_Replications(matrix)
}
\arguments{
\item{matrix}{Any matrix}
}
\value{
Returns a matrix of replications for each treatment.
}
\description{
Replications for each treatments
}
\examples{
library(SudokuDesigns)
mat11<-matrix(c(1,2,3,4,1,3,6,2,8,1,8,3),nrow=4,byrow=TRUE)
mat11
Check_Replications(mat11)
}
