% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Sudoku_I.R
\name{Get_Sudoku_I}
\alias{Get_Sudoku_I}
\title{Complete/Incomplete Sudoku Designs for Even Number, v}
\usage{
Get_Sudoku_I(v, type = "complete")
}
\arguments{
\item{v}{Please enter an number, v}

\item{type}{Please choose type as "complete" or "incomplete". Default is "complete".}
}
\value{
For a given v, this function will provide the Sudoku design and its parameters, Region matrix, C matrix, eigenvalues (EVs) and canonical efficiency factor (CEF).
}
\description{
To obtain complete as well as incomplete Sudoku designs for an even number one can use this function. The generated designs are a new series of Sudoku designs.
}
\examples{
library(SudokuDesigns)
Get_Sudoku_I(10)
}
