% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Sudoku_II.R
\name{Get_Sudoku_II}
\alias{Get_Sudoku_II}
\title{Incomplete Sudoku designs for v = nC2 where n (>=5) is an odd number}
\usage{
Get_Sudoku_II(v)
}
\arguments{
\item{v}{Provide v = nC2 where n (>=5) is an odd number}
}
\value{
It returns an incomplete Gerechte design along with its parameters, region matrix, C matrix, eigenvalues (EVs) and canonical efficiency factor (CEF).
}
\description{
Generated designs with less number of regions with quite high canonical efficiency factors.
}
\examples{
library(SudokuDesigns)
Get_Sudoku_II(10)
}
